claimler eklendi

+++ Entities/DTOs/VendorProductDetailDto.cs
@@ -9,8 +9,8 @@ namespace Entities.DTOs
 {
     public class VendorProductDetailDto:IDto
     {
-        public decimal Price { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int Price { get; set; }
+        public int UnitPrice { get; set; }
         public int Quantity { get; set; }
         public string Description { get; set; }
         public int CategoryID { get; set; }
